move some parts to common package
