+++ app/main.py
@@ -1,24 +1,26 @@
 from fastapi import FastAPI
-from loguru import logger
 
-from app.config import APP_VERSION, DEBUG
+from app.common.config import APP_VERSION, DEBUG
+from app.common.events import (
+    create_shutdown_events_handler,
+    create_startup_events_handler,
+)
 from app.endpoints.api.computers import router as computers_router
 from app.endpoints.websockets import router as ws_router
 
-app = FastAPI(title="Mirumon service", version=APP_VERSION, debug=DEBUG)
 
-app.include_router(router=computers_router)
-app.include_router(router=ws_router)
+def get_application() -> FastAPI:
+    application = FastAPI(title="Mirumon service", version=APP_VERSION, debug=DEBUG)
+
+    application.include_router(router=computers_router)
+    application.include_router(router=ws_router)
+
+    application.add_event_handler("startup", create_startup_events_handler(application))
+    application.add_event_handler(
+        "shutdown", create_shutdown_events_handler(application)
+    )
+
+    return application
 
 
-@logger.catch
-async def startup() -> None:
-    logger.info("startup app")
-
-
-async def shutdown() -> None:
-    logger.info("shutdown app")
-
-
-app.add_event_handler("startup", startup)
-app.add_event_handler("shutdown", shutdown)
+app = get_application()
